requirements.txt
# =============================================================================
# trip-planner 项目依赖汇总
# =============================================================================
# 各子模块保留自己的 requirements.txt（用于独立安装），本文件作为全局总览。
#
# 安装方式:
#   后端:       cd backend    && pip install -r requirements.txt
#   MCP Server: cd mcp-server && pip install -r requirements.txt
#   前端:       cd frontend   && npm install    （由 package.json 管理）
#
# Python 环境要求: 3.10+    Node.js 环境要求: 18+
# =============================================================================


# -----------------------------------------------------------------------------
# [backend]  FastAPI 服务 + LangGraph 多智能体框架
# 详见: backend/requirements.txt
# -----------------------------------------------------------------------------

# LangGraph / LangChain 核心框架
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0         # ChatOpenAI，兼容 OpenAI 格式接口（DeepSeek 等）
langchain-mcp-adapters>=0.1.0   # MCP stdio 工具适配层，连接 mcp-server 子进程

# FastAPI Web 服务
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0.0
pydantic-settings>=2.0.0        # 从根目录 .env 自动绑定 Settings 字段

# 工具库
httpx>=0.27.0                   # 异步 HTTP 客户端
python-dotenv>=1.0.0            # 加载根目录 .env
python-multipart>=0.0.9         # CORS / multipart form 支持
loguru>=0.7.0                   # 结构化日志
python-dateutil>=2.8.2          # 日期解析工具
redis[asyncio]>=5.0.0           # 异步 Redis 客户端（行程计划缓存，REDIS_URL 为空时自动禁用）

# 测试框架
pytest>=8.0.0                   # 单元测试框架
pytest-asyncio>=0.24.0          # asyncio 测试支持
pytest-cov>=5.0.0               # 覆盖率统计


# -----------------------------------------------------------------------------
# [mcp-server]  独立 Amap MCP Server（FastMCP 实现，stdio 子进程）
# 详见: mcp-server/requirements.txt
# -----------------------------------------------------------------------------

mcp[cli]>=1.0.0                 # FastMCP 框架，提供 @mcp.tool() 装饰器与 stdio transport
# httpx>=0.27.0                 # 与 backend 共享，已在上方声明
# python-dotenv>=1.0.0          # 同上


# -----------------------------------------------------------------------------
# [frontend]  Vue 3 前端（Node.js，由 frontend/package.json 管理）
# 以下仅作版本记录，不通过 pip 安装
# -----------------------------------------------------------------------------

# vue                      ^3.5.13   核心框架
# vue-router               ^4.5.0    路由管理
# ant-design-vue           ^4.2.6    UI 组件库
# axios                    ^1.7.9    HTTP 客户端，与 backend 通信
# @amap/amap-jsapi-loader  ^1.0.1    高德地图 JS API 加载器（地图渲染）
# html2canvas              ^1.4.1    行程页面截图
# jspdf                    ^3.0.3    PDF 导出
# vite                     ^6.0.7    构建工具
# @vitejs/plugin-vue       ^5.2.1    Vue SFC 编译插件
# typescript               ^5.7.3
# vue-tsc                  ^2.2.0    TypeScript 类型检查

